pyproject.toml
[tool.poetry]
name = "selective-imitation-learning"
version = "0.1.0"
description = ""
authors = ["Max Taylor-Davies <maxtaylordavies@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.5.2"
gymnasium = "^0.29.1"
torch = "^2.2.1"
matplotlib = "^3.8.3"
tqdm = "^4.66.2"
stable-baselines3 = "^2.2.1"
rich = "^13.7.1"
seaborn = "^0.13.2"
imitation = "^1.0.0"
jax = "^0.4.25"
numpyro = "^0.14.0"
optax = "^0.2.2"
equinox = "^0.11.3"
plotly = "^5.20.0"
moviepy = "^1.0.3"
array2gif = "^1.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
